Make charts responsible for loading queries
This feels... wrong, since charts shouldn't have to know about the API,
but it neatly solves the problem of the first chart not loading when the
initial page is the list of slider settings.

It also eliminates the ability to efficiently display multiple charts on
one page, but that feature isn't needed at this time.